Learned the mediator pattern. It is like an inverse observer.
I am seeing that a lot of these patterns have the same or a similar UML diagram, it is just that they
have different contexts. Knowing when to use a given pattern might be tricky without experience, but
the biggest takeaway so far has been the use of interfaces and abstractions to decouple code.