Done with the MVC pattern...and all the major design patterns!
That is not to say I have truly learned these, as knowing how to apply these patterns is going to be a lot more difficulty than simply knowing how to make one. Regardless, that will come with experience and my next goal is to learn the MERN stack and test driven development.